Changed type of stride variables from 32-bit integers to 64-bit integers to avoid problems with splines > 2GB.

+++ src/multi_bspline_eval_sse_c_impl.h
@@ -78,7 +78,7 @@ eval_multi_UBspline_1d_c (multi_UBspline_1d_c *spline,
   a[2]  = (Af[ 8]*tpx[0] + Af[ 9]*tpx[1] + Af[10]*tpx[2] + Af[11]*tpx[3]);
   a[3]  = (Af[12]*tpx[0] + Af[13]*tpx[1] + Af[14]*tpx[2] + Af[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) 
     vals[n]  = 0.0;
@@ -116,7 +116,7 @@ eval_multi_UBspline_1d_c_vg (multi_UBspline_1d_c *spline,
   da[2] = (dAf[ 8]*tpx[0] + dAf[ 9]*tpx[1] + dAf[10]*tpx[2] + dAf[11]*tpx[3]);
   da[3] = (dAf[12]*tpx[0] + dAf[13]*tpx[1] + dAf[14]*tpx[2] + dAf[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n]  = 0.0;
@@ -166,7 +166,7 @@ eval_multi_UBspline_1d_c_vgl (multi_UBspline_1d_c *spline,
   d2a[2] = (d2Af[ 8]*tpx[0] + d2Af[ 9]*tpx[1] + d2Af[10]*tpx[2] + d2Af[11]*tpx[3]);
   d2a[3] = (d2Af[12]*tpx[0] + d2Af[13]*tpx[1] + d2Af[14]*tpx[2] + d2Af[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n]  = 0.0;
@@ -256,8 +256,8 @@ eval_multi_UBspline_2d_c (multi_UBspline_2d_c *spline,
   // Zero out values;
   for (int n=0; n<Nm; n++)     mvals[n] = _mm_setzero_ps();
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   // Main compute loop
   __m128 ab;
@@ -337,8 +337,8 @@ eval_multi_UBspline_2d_c_vg (multi_UBspline_2d_c *spline,
   for (int n=0; n<Nm; n++)     mvals[n] = _mm_setzero_ps();
   for (int n=0; n<2*Nm; n++)   mgrad[n] = _mm_setzero_ps();
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   // Main compute loop
   __m128 ab, dab[2];
@@ -441,8 +441,8 @@ eval_multi_UBspline_2d_c_vgl (multi_UBspline_2d_c *spline,
   for (int n=0; n<2*Nm; n++)   mgrad[n] = _mm_setzero_ps();
   for (int n=0; n<2*Nm; n++)   mlapl[n] = _mm_setzero_ps();
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   // Main compute loop
   __m128 ab, dab[2], d2ab[2];
@@ -579,8 +579,8 @@ eval_multi_UBspline_2d_c_vgh (multi_UBspline_2d_c *spline,
   for (int n=0; n<2*Nm; n++)   mgrad[n] = _mm_setzero_ps();
   for (int n=0; n<3*Nm; n++)   mhess[n] = _mm_setzero_ps();
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   // Main compute loop
   __m128 ab, dab[2], d2ab[3];
@@ -700,9 +700,9 @@ eval_multi_UBspline_3d_c (multi_UBspline_3d_c *spline,
   for (int n=0; n<Nm; n++)     mvals[n] = _mm_setzero_ps();
 
   // Main compute loop
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   __m128 abc;
   for (int i=0; i<4; i++)
@@ -752,9 +752,9 @@ eval_multi_UBspline_3d_c_vg (multi_UBspline_3d_c *spline,
   int iy = ((int *)&ixiyiz)[2];
   int iz = ((int *)&ixiyiz)[1];
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   // Now compute the vectors:
   // tpx = [t_x^3 t_x^2 t_x 1]
@@ -884,9 +884,9 @@ eval_multi_UBspline_3d_c_vgl (multi_UBspline_3d_c *spline,
   int iy = ((int *)&ixiyiz)[2];
   int iz = ((int *)&ixiyiz)[1];
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   // Now compute the vectors:
   // tpx = [t_x^3 t_x^2 t_x 1]
@@ -1035,9 +1035,9 @@ eval_multi_UBspline_3d_c_vgh (multi_UBspline_3d_c *spline,
   int iy = ((int *)&ixiyiz)[2];
   int iz = ((int *)&ixiyiz)[1];
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   // Now compute the vectors:
   // tpx = [t_x^3 t_x^2 t_x 1]
