+++ src/multi_bspline_eval_std_c_impl.h
@@ -52,7 +52,7 @@ eval_multi_UBspline_1d_c (multi_UBspline_1d_c *spline,
   a[2]  = (Af[ 8]*tpx[0] + Af[ 9]*tpx[1] + Af[10]*tpx[2] + Af[11]*tpx[3]);
   a[3]  = (Af[12]*tpx[0] + Af[13]*tpx[1] + Af[14]*tpx[2] + Af[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) 
     vals[n]  = 0.0;
@@ -90,7 +90,7 @@ eval_multi_UBspline_1d_c_vg (multi_UBspline_1d_c *spline,
   da[2] = (dAf[ 8]*tpx[0] + dAf[ 9]*tpx[1] + dAf[10]*tpx[2] + dAf[11]*tpx[3]);
   da[3] = (dAf[12]*tpx[0] + dAf[13]*tpx[1] + dAf[14]*tpx[2] + dAf[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n]  = 0.0;
@@ -140,7 +140,7 @@ eval_multi_UBspline_1d_c_vgl (multi_UBspline_1d_c *spline,
   d2a[2] = (d2Af[ 8]*tpx[0] + d2Af[ 9]*tpx[1] + d2Af[10]*tpx[2] + d2Af[11]*tpx[3]);
   d2a[3] = (d2Af[12]*tpx[0] + d2Af[13]*tpx[1] + d2Af[14]*tpx[2] + d2Af[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n]  = 0.0;
@@ -207,8 +207,8 @@ eval_multi_UBspline_2d_c (multi_UBspline_2d_c *spline,
   b[2] = (Af[ 8]*tpy[0] + Af[ 9]*tpy[1] + Af[10]*tpy[2] + Af[11]*tpy[3]);
   b[3] = (Af[12]*tpy[0] + Af[13]*tpy[1] + Af[14]*tpy[2] + Af[15]*tpy[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   for (int n=0; n<spline->num_splines; n++)
     vals[n] = 0.0;
@@ -260,8 +260,8 @@ eval_multi_UBspline_2d_c_vg (multi_UBspline_2d_c *spline,
   db[2] = (dAf[ 8]*tpy[0] + dAf[ 9]*tpy[1] + dAf[10]*tpy[2] + dAf[11]*tpy[3]);
   db[3] = (dAf[12]*tpy[0] + dAf[13]*tpy[1] + dAf[14]*tpy[2] + dAf[15]*tpy[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n] = 0.0;
@@ -337,8 +337,8 @@ eval_multi_UBspline_2d_c_vgl (multi_UBspline_2d_c *spline,
   d2b[2] = (d2Af[ 8]*tpy[0] + d2Af[ 9]*tpy[1] + d2Af[10]*tpy[2] + d2Af[11]*tpy[3]);
   d2b[3] = (d2Af[12]*tpy[0] + d2Af[13]*tpy[1] + d2Af[14]*tpy[2] + d2Af[15]*tpy[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   complex_float lapl2[2*spline->num_splines];
   for (int n=0; n<spline->num_splines; n++) {
@@ -423,8 +423,8 @@ eval_multi_UBspline_2d_c_vgh (multi_UBspline_2d_c *spline,
   d2b[2] = (d2Af[ 8]*tpy[0] + d2Af[ 9]*tpy[1] + d2Af[10]*tpy[2] + d2Af[11]*tpy[3]);
   d2b[3] = (d2Af[12]*tpy[0] + d2Af[13]*tpy[1] + d2Af[14]*tpy[2] + d2Af[15]*tpy[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n] = 0.0;
@@ -508,9 +508,9 @@ eval_multi_UBspline_3d_c (multi_UBspline_3d_c *spline,
   c[2] = (Af[ 8]*tpz[0] + Af[ 9]*tpz[1] + Af[10]*tpz[2] + Af[11]*tpz[3]);
   c[3] = (Af[12]*tpz[0] + Af[13]*tpz[1] + Af[14]*tpz[2] + Af[15]*tpz[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   for (int n=0; n<spline->num_splines; n++)
     vals[n] = 0.0;
@@ -577,9 +577,9 @@ eval_multi_UBspline_3d_c_vg (multi_UBspline_3d_c *spline,
   dc[2] = (dAf[ 8]*tpz[0] + dAf[ 9]*tpz[1] + dAf[10]*tpz[2] + dAf[11]*tpz[3]);
   dc[3] = (dAf[12]*tpz[0] + dAf[13]*tpz[1] + dAf[14]*tpz[2] + dAf[15]*tpz[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n] = 0.0;
@@ -680,9 +680,9 @@ eval_multi_UBspline_3d_c_vgl (multi_UBspline_3d_c *spline,
   d2c[2] = (d2Af[ 8]*tpz[0] + d2Af[ 9]*tpz[1] + d2Af[10]*tpz[2] + d2Af[11]*tpz[3]);
   d2c[3] = (d2Af[12]*tpz[0] + d2Af[13]*tpz[1] + d2Af[14]*tpz[2] + d2Af[15]*tpz[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   complex_float lapl3[3*spline->num_splines];
   for (int n=0; n<spline->num_splines; n++) {
@@ -795,9 +795,9 @@ eval_multi_UBspline_3d_c_vgh (multi_UBspline_3d_c *spline,
   d2c[2] = (d2Af[10]*tpz[2] + d2Af[11]*tpz[3]);
   d2c[3] = (d2Af[14]*tpz[2] + d2Af[15]*tpz[3]);
 
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n] = 0.0;
